Refactor the names of classes
